Use token from last collection item for next call
After a quick modification to grab and save "collection_seed"
and "data" dictionaries to disk I discovered that the
data["last_token"] field is not set properly by the server when the
"count" requested is > 45. When more than 45 items are returned, the
"last_token" returned as if for the 45th item, not the 100th item or
whatever "count" is requested. Thus you'd get responses like:
request 1: items 0-100
request 2: items 46-146
request 3: items 91-191
and so on with lots of duplicated entries between requests.

With this change the code now uses the token from the final item
in each response list, thus request 2 would start from 101.